+++ components/solana-section.tsx
@@ -1,18 +1,19 @@
 import { encode } from '@/lib/utils';
+//@ts-ignore
 import { useWallet } from '@coin98t/wallet-adapter-react';
-import bs58 from 'bs58';
 import {
   Connection,
   Keypair,
   LAMPORTS_PER_SOL,
   PublicKey,
   SystemProgram,
-  Transaction as TransactionSolana,
   TransactionInstruction,
   TransactionMessage,
+  Transaction as TransactionSolana,
   VersionedTransaction,
 } from '@solana/web3.js';
-import { useEffect, useState } from 'react';
+import bs58 from 'bs58';
+import { useState } from 'react';
 import CardMethod from './ui/card-method';
 import CustomButton from './ui/custom-button';
 import ResultTxt from './ui/resultTxt';
@@ -22,7 +23,7 @@ const ContentSolana = () => {
   const connection = new Connection(
     'https://rough-white-cloud.solana-mainnet.discover.quiknode.pro/917d316c92433f9d91a7c0c16299df93e2883054/',
   );
-  const lamports = 0.001;
+  const lamports = 0.0001;
   let recipientAddress = '41J69vTXwFyjzBHahVhzN32Fty77DXDCa7EG8maTyYGy';
 
   // Hook
@@ -35,8 +36,7 @@ const ContentSolana = () => {
 
   const handleSignMessage = async () => {
     try {
-      const response = await signMessage?.(new TextEncoder().encode('ChiPoPo'));
-
+      const response: any = await signMessage?.(new TextEncoder().encode('ChiPoPo'));
       setResultMessage(Buffer.from(response.data as any).toString('hex'));
     } catch (error) {
       console.log(error);
@@ -54,13 +54,15 @@ const ContentSolana = () => {
         new TransactionInstruction({
           data: Buffer.from('Hello, from the Coin98 Wallet Adapter example app!'),
           keys: [],
-          programId: new PublicKey('MemoSq4gqABAXKb96qnH8TysNcWxMyWCqXgDLGmfcHr'),
+          programId: new PublicKey(publicKey as PublicKey),
         }),
       );
-
-      const resSig = await signTransaction?.(transaction);
-      console.log('verifySignatures', resSig?.verifySignatures());
-      setResultSignTrans(encode(transaction.signature as any));
+      const resSig: any = await signTransaction?.(transaction);
+      if (transaction?.verifySignatures) {
+        console.log('verifySignatures', transaction.verifySignatures());
+      }
+      console.log({ resSig });
+      setResultSignTrans(encode(resSig.signature));
     } catch (error) {
       console.log('error sigTransaction', error);
     }
@@ -68,9 +70,9 @@ const ContentSolana = () => {
 
   const handleSendToken = async () => {
     const kp = Keypair.fromSecretKey(
-      bs58.decode('37UL1qZ3yJBBpxErzUYVwXRMaL7JayBWYxpRZZPXiiBCQSGpykiaVyisw7tuKrjBc5GyP8MY852B52vyQDTLpkZE'),
+      //private key
+      bs58.decode(''),
     );
-
     try {
       if (!publicKey) throw new Error();
 
@@ -94,7 +96,12 @@ const ContentSolana = () => {
         skipPreflight: false,
         signers: [{ publicKey: kp.publicKey, secretKey: kp.secretKey }],
       });
-
+      // console.log({ resSend });
+      // const rawTransaction = transaction.serialize();
+      // console.log({ rawTransaction });
+      // const res = await connection.sendRawTransaction(rawTransaction);
+      // console.log({ res });
+      //@ts-ignore //done
       setResultSend((resSend.data as string) || (resSend.error as string));
     } catch (error) {
       console.log('error sendToken', error);
@@ -123,7 +130,7 @@ const ContentSolana = () => {
     // make a versioned transaction
     const transactionV0 = new VersionedTransaction(messageV0);
 
-    // const response = await window.coin98.sol.request({
+    // const response = await window.saros.request({
     //   method: 'sol_sign',
     //   params: [transactionV0],
     // });
@@ -144,7 +151,8 @@ const ContentSolana = () => {
       skipPreflight: false,
       signers: [],
     });
-
+    //done
+    //@ts-ignore
     setResultSendVerTrans((resSend.data as string) || (resSend.error as string));
   };
 
@@ -159,7 +167,7 @@ const ContentSolana = () => {
         new TransactionInstruction({
           data: Buffer.from('Hello, from the Coin98 Wallet Adapter example app!'),
           keys: [],
-          programId: new PublicKey('MemoSq4gqABAXKb96qnH8TysNcWxMyWCqXgDLGmfcHr'),
+          programId: new PublicKey(publicKey as PublicKey),
         }),
       );
 
@@ -254,11 +262,12 @@ const ContentSolana = () => {
         <CustomButton onClick={() => handleSignTransaction()} title="Sign" className="mt-6" />
         {resultSignTrans && <ResultTxt txt={resultSignTrans} />}
       </CardMethod>
-
-      <CardMethod title="Send Token Bybit">
-        <CustomButton onClick={() => handleSendTransactionSol()} title="Send" className="mt-6" />
-        {resultSignTrans && <ResultTxt txt={resultSignTrans} />}
-      </CardMethod>
+      {window?.bybitWallet && (
+        <CardMethod title="Send Token Bybit">
+          <CustomButton onClick={() => handleSendTransactionSol()} title="Send" className="mt-6" />
+          {resultSignTrans && <ResultTxt txt={resultSignTrans} />}
+        </CardMethod>
+      )}
     </div>
   );
 };